\name{Barnard}
\alias{Barnard}
\alias{Barnard-package}
\docType{package}

\title{
Barnard's Unconditional Test
}

\description{
This package implements the \code{barnard.test} function for performing Barnard's unconditional test of superiority.
This is a more powerful alternative of Fisher's exact test for 2x2 contingency tables. The test, in its current implementation, uses Wald statistics as a measure of difference between two binomial proportions.
}

\author{
\enumerate{
\item Kamil Erguler, Post-doctoral Fellow, EEWRC, The Cyprus Institute \email{k.erguler@cyi.ac.cy}
}
}

\references{
\enumerate{
\item Barnard, G.A. (1945) A new test for 2x2 tables. \emph{Nature}, 156:177.
\item Barnard, G.A. (1947) Significance tests for 2x2 tables. \emph{Biometrika}, 34:123-138.
}
}

\keyword{nonparametric}
\keyword{htest}


