\name{barnardw.test}
\alias{barnardw.test}
\title{ Barnard's Unconditional Test with Wald Statistics (obsolete)}
\description{
Previous version of Barnard's unconditional test for superiority which
used Z-statistic with pooled variance for the difference between two binomial
proportions in a 2x2 contingency table. Please use the 'barnard.test' instead.
}
